feat() mainactivity deleted, tabview activity renamed, and user sign in checked
